ansible-silverblue-flatpaks.jinja2: Changes so this can be rerun
If you re-run the script that results from ansible-silverblue-flatpaks.jinja2
now, it will error-out. I'm adding some conditionals so that it will not bomb
out on subsequent runs.
This will allow users to remove the file that gets laid down-on firstboot, update
their OS, and then have this script run again automatically.